feat: rootNavigation and notifications
