+++ package.json
@@ -23,7 +23,8 @@
     "react-native-pager-view": "^6.2.0",
     "react-native-paper": "^5.7.0",
     "react-native-paper-tabs": "^0.7.0",
-    "react-native-reanimated": "~2.14.4"
+    "react-native-reanimated": "~2.14.4",
+    "react-navigation-header-buttons": "^10.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
