+++ src/components/Notifications.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect, useRef } from 'react'
 import { Text, View, Button, Platform } from 'react-native'
 import * as Device from 'expo-device'
-import * as Notifications from 'expo-notifications'
+import * as ExpoNotifications from 'expo-notifications'
 
-Notifications.setNotificationHandler({
+ExpoNotifications.setNotificationHandler({
   handleNotification: async () => ({
     shouldShowAlert: true,
     shouldPlaySound: false,
@@ -35,26 +35,27 @@ async function sendPushNotification(expoPushToken) {
 async function registerForPushNotificationsAsync() {
   let token
   if (Device.isDevice) {
-    const { status: existingStatus } = await Notifications.getPermissionsAsync()
+    const { status: existingStatus } =
+      await ExpoNotifications.getPermissionsAsync()
     let finalStatus = existingStatus
     if (existingStatus !== 'granted') {
-      const { status } = await Notifications.requestPermissionsAsync()
+      const { status } = await ExpoNotifications.requestPermissionsAsync()
       finalStatus = status
     }
     if (finalStatus !== 'granted') {
       alert('Failed to get push token for push notification!')
       return
     }
-    token = (await Notifications.getExpoPushTokenAsync()).data
+    token = (await ExpoNotifications.getExpoPushTokenAsync()).data
     console.log(token)
   } else {
     alert('Must use physical device for Push Notifications')
   }
 
   if (Platform.OS === 'android') {
-    Notifications.setNotificationChannelAsync('default', {
+    ExpoNotifications.setNotificationChannelAsync('default', {
       name: 'default',
-      importance: Notifications.AndroidImportance.MAX,
+      importance: ExpoNotifications.AndroidImportance.MAX,
       vibrationPattern: [0, 250, 250, 250],
       lightColor: '#FF231F7C',
     })
@@ -63,7 +64,7 @@ async function registerForPushNotificationsAsync() {
   return token
 }
 
-export default function Chat() {
+export default function Notifications() {
   const [expoPushToken, setExpoPushToken] = useState('')
   const [notification, setNotification] = useState(false)
   const notificationListener = useRef()
@@ -73,18 +74,20 @@ export default function Chat() {
     registerForPushNotificationsAsync().then((token) => setExpoPushToken(token))
 
     notificationListener.current =
-      Notifications.addNotificationReceivedListener((notification) => {
+      ExpoNotifications.addNotificationReceivedListener((notification) => {
         setNotification(notification)
       })
 
     responseListener.current =
-      Notifications.addNotificationResponseReceivedListener((response) => {
+      ExpoNotifications.addNotificationResponseReceivedListener((response) => {
         console.log(response)
       })
 
     return () => {
-      Notifications.removeNotificationSubscription(notificationListener.current)
-      Notifications.removeNotificationSubscription(responseListener.current)
+      ExpoNotifications.removeNotificationSubscription(
+        notificationListener.current
+      )
+      ExpoNotifications.removeNotificationSubscription(responseListener.current)
     }
   }, [])
