+++ src/components/TenderRequests.tsx
@@ -4,12 +4,13 @@ import {
   Button,
   Card,
   Checkbox,
+  IconButton,
   List,
   Searchbar,
   Text,
   Title,
 } from 'react-native-paper'
-import { ScrollView, StyleSheet, View } from 'react-native'
+import { Alert, ScrollView, StyleSheet, View } from 'react-native'
 import { StatusBar } from 'expo-status-bar'
 import tenderRequests from '../data/tenderRequests'
 import MaterialCommunityIcons from 'react-native-vector-icons/MaterialCommunityIcons'
@@ -25,11 +26,6 @@ const ChevronRight = () => (
 const TenderRequests = ({ navigation }) => {
   const [searchQuery, setSearchQuery] = React.useState('')
   const [open, setOpen] = React.useState(true)
-  const [checkboxStatus, setCheckboxStatus] = React.useState({
-    favoriter: false,
-    öppna: false,
-    tilldelade: false,
-  })
 
   const filteredRequests = tenderRequests.filter((request) =>
     request.title.toLowerCase().includes(searchQuery.toLowerCase())
